Revert "Single article page controller, model and layout"
