+++ .gitignore
@@ -1,4 +1,3 @@
-/.idea
 
 ###> symfony/framework-bundle ###
 /.env.local
