+++ composer.json
@@ -7,13 +7,6 @@
         "php": ">=7.4",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "composer/package-versions-deprecated": "^1.11",
-        "doctrine/common": "~2",
-        "doctrine/doctrine-bundle": "^2.1",
-        "doctrine/doctrine-migrations-bundle": "^3.0",
-        "doctrine/orm": "^2.7",
-        "doctrine/persistence": "~1.3",
-        "stof/doctrine-extensions-bundle": "^1.4",
         "symfony/asset": "5.1.*",
         "symfony/console": "5.1.*",
         "symfony/dotenv": "5.1.*",
@@ -25,10 +18,8 @@
         "twig/twig": "^2.12|^3.0"
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.3",
         "friendsofphp/php-cs-fixer": "^2.16",
-        "fzaninotto/faker": "^1.9",
-        "symfony/maker-bundle": "^1.21"
+        "fzaninotto/faker": "^1.9"
     },
     "config": {
         "optimize-autoloader": true,
