--- config/packages/doctrine.yaml
@@ -1,15 +1,0 @@
-doctrine:
-    dbal:
-        url: '%env(resolve:DATABASE_URL)%'
-        server_version: '8.0'
-    orm:
-        auto_generate_proxy_classes: true
-        naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
-        auto_mapping: true
-        mappings:
-            App:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/src/Entity'
-                prefix: 'App\Entity'
-                alias: App
