--- config/packages/stof_doctrine_extensions.yaml
@@ -1,4 +1,0 @@
-# Read the documentation: https://symfony.com/doc/current/bundles/StofDoctrineExtensionsBundle/index.html
-# See the official DoctrineExtensions documentation for more details: https://github.com/Atlantic18/DoctrineExtensions/tree/master/doc/
-stof_doctrine_extensions:
-    default_locale: en_US
