--- public/css/single.css
@@ -1,9 +1,0 @@
-.single-article, .single-category {
-    padding: 0 30px;
-}
-.single-article .title-wrap, .single-category .title-wrap {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: space-between;
-}
