--- src/DataFixtures/AbstractFixture.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\DataFixtures;
-
-use Doctrine\Bundle\FixturesBundle\Fixture;
-use Faker\Factory;
-use Faker\Generator;
-
-abstract class AbstractFixture extends Fixture
-{
-    protected Generator $faker;
-
-    public function __construct()
-    {
-        $this->faker = Factory::create();
-    }
-}
