--- src/Exception/ArticleBodyCannotBeEmptyException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exception;
-
-final class ArticleBodyCannotBeEmptyException extends \DomainException
-{
-}
