--- src/Service/HomePageArticlesDBProvider.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Service;
-
-use App\Collection\HomePageArticles;
-use App\Entity\Article;
-use App\Repository\ArticleRepository;
-
-final class HomePageArticlesDBProvider implements HomePageArticlesProviderInterface
-{
-    private ArticleRepository $articleRepository;
-
-    public function __construct(ArticleRepository $articleRepository)
-    {
-        $this->articleRepository = $articleRepository;
-    }
-
-    public function getList(): HomePageArticles
-    {
-        $articles = $this->articleRepository->getLatestPublished();
-
-        $viewModels = \array_map(fn (Article $article) => $article->getHomePageArticle(), $articles);
-
-        return new HomePageArticles(...$viewModels);
-    }
-}
