+++ symfony.lock
@@ -1,10 +1,4 @@
 {
-    "behat/transliterator": {
-        "version": "v1.3.0"
-    },
-    "composer/package-versions-deprecated": {
-        "version": "1.11.99"
-    },
     "composer/semver": {
         "version": "1.5.1"
     },
@@ -23,87 +17,8 @@
             "config/routes/annotations.yaml"
         ]
     },
-    "doctrine/cache": {
-        "version": "1.10.2"
-    },
-    "doctrine/collections": {
-        "version": "1.6.7"
-    },
-    "doctrine/common": {
-        "version": "3.0.2"
-    },
-    "doctrine/data-fixtures": {
-        "version": "1.4.4"
-    },
-    "doctrine/dbal": {
-        "version": "2.10.3"
-    },
-    "doctrine/doctrine-bundle": {
-        "version": "2.0",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "2.0",
-            "ref": "a9f2463b9f73efe74482f831f03a204a41328555"
-        },
-        "files": [
-            "config/packages/doctrine.yaml",
-            "config/packages/prod/doctrine.yaml",
-            "src/Entity/.gitignore",
-            "src/Repository/.gitignore"
-        ]
-    },
-    "doctrine/doctrine-fixtures-bundle": {
-        "version": "3.0",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "3.0",
-            "ref": "e5b542d4ef47d8a003c91beb35650c76907f7e53"
-        },
-        "files": [
-            "src/DataFixtures/AppFixtures.php"
-        ]
-    },
-    "doctrine/doctrine-migrations-bundle": {
-        "version": "2.2",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "2.2",
-            "ref": "baaa439e3e3179e69e3da84b671f0a3e4a2f56ad"
-        },
-        "files": [
-            "config/packages/doctrine_migrations.yaml",
-            "migrations/.gitignore"
-        ]
-    },
-    "doctrine/event-manager": {
-        "version": "1.1.1"
-    },
-    "doctrine/inflector": {
-        "version": "1.4.3"
-    },
-    "doctrine/instantiator": {
-        "version": "1.3.1"
-    },
     "doctrine/lexer": {
         "version": "1.2.1"
-    },
-    "doctrine/migrations": {
-        "version": "3.0.1"
-    },
-    "doctrine/orm": {
-        "version": "v2.7.3"
-    },
-    "doctrine/persistence": {
-        "version": "2.0.0"
-    },
-    "doctrine/reflection": {
-        "version": "1.2.1"
-    },
-    "doctrine/sql-formatter": {
-        "version": "1.1.1"
     },
     "friendsofphp/php-cs-fixer": {
         "version": "2.2",
@@ -119,24 +34,6 @@
     },
     "fzaninotto/faker": {
         "version": "v1.9.1"
-    },
-    "gedmo/doctrine-extensions": {
-        "version": "v2.4.42"
-    },
-    "laminas/laminas-code": {
-        "version": "3.4.1"
-    },
-    "laminas/laminas-eventmanager": {
-        "version": "3.3.0"
-    },
-    "laminas/laminas-zendframework-bridge": {
-        "version": "1.1.0"
-    },
-    "nikic/php-parser": {
-        "version": "v4.9.1"
-    },
-    "ocramius/proxy-manager": {
-        "version": "2.8.0"
     },
     "php": {
         "version": "7.4"
@@ -155,18 +52,6 @@
     },
     "psr/log": {
         "version": "1.1.3"
-    },
-    "stof/doctrine-extensions-bundle": {
-        "version": "1.2",
-        "recipe": {
-            "repo": "github.com/symfony/recipes-contrib",
-            "branch": "master",
-            "version": "1.2",
-            "ref": "6c1ceb662f8997085f739cd089bfbef67f245983"
-        },
-        "files": [
-            "config/packages/stof_doctrine_extensions.yaml"
-        ]
     },
     "symfony/asset": {
         "version": "v5.1.3"
@@ -197,9 +82,6 @@
     },
     "symfony/deprecation-contracts": {
         "version": "v2.1.3"
-    },
-    "symfony/doctrine-bridge": {
-        "version": "v5.1.5"
     },
     "symfony/dotenv": {
         "version": "v5.1.3"
@@ -256,20 +138,8 @@
     "symfony/http-kernel": {
         "version": "v5.1.3"
     },
-    "symfony/maker-bundle": {
-        "version": "1.0",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "master",
-            "version": "1.0",
-            "ref": "fadbfe33303a76e25cb63401050439aa9b1a9c7f"
-        }
-    },
     "symfony/options-resolver": {
         "version": "v5.1.3"
-    },
-    "symfony/orm-pack": {
-        "version": "v2.0.0"
     },
     "symfony/polyfill-intl-grapheme": {
         "version": "v1.18.1"
@@ -349,8 +219,5 @@
     },
     "twig/twig": {
         "version": "v3.0.5"
-    },
-    "webimpress/safe-writer": {
-        "version": "2.1.0"
     }
 }
