--- templates/404.html.twig
@@ -1,8 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body %}
-    <section class="not-found">
-        <h1>Page not found</h1>
-        <p>You can visit the <a href="{{ path('app_home') }}">homepage</a> to see latest articles.</p>
-    </section>
-{% endblock %}
