--- templates/articles/single.html.twig
@@ -1,21 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block styles %}
-    <link rel="stylesheet" href="{{ asset('css/single.css') }}">
-{% endblock %}
-
-{% block body %}
-    <article class="single-article">
-        <div class="title-wrap">
-            <div class="title">
-                <a href="{{ path('app_category_page', {slug: article.category.slug}) }}" class="category-link">{{ article.category.title }}</a>
-                <h1>{{ article.title }}</h1>
-            </div>
-            <div class="date">
-                <time>{{ article.publicationDate | date('d-m-Y') }}</time>
-            </div>
-        </div>
-        <p>{{ article.body }}</p>
-        <div class="clearfix"></div>
-    </article>
-{% endblock %}
