--- templates/categories/single.html.twig
@@ -1,32 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block styles %}
-    <link rel="stylesheet" href="{{ asset('css/single.css') }}">
-{% endblock %}
-
-{% block body %}
-    <article class="single-category">
-        <div class="title-wrap">
-            <div class="title">
-                <h1>{{ category.title }}</h1>
-            </div>
-        </div>
-        <div class="row posts">
-            {% for article in category.articles %}
-                <div class="col-3">
-                    <article class="post top-post">
-                        <a href="{{ path('app_article_page', {id: article.id}) }}">
-                            <img src="{{ asset(article.image) }}" alt="" class="img-fluid">
-                        </a>
-                        <a href="{{ path('app_article_page', {id: article.id}) }}" class="post-link">
-                            <h3>{{ article.title }}</h3>
-                        </a>
-                        <time>{{ article.publicationDate | date('d-m-Y') }}</time>
-                        <p>{{ article.shortDescription }}</p>
-                    </article>
-                </div>
-            {% endfor %}
-        </div>
-        <div class="clearfix"></div>
-    </article>
-{% endblock %}
