+++ templates/home/index.html.twig
@@ -9,11 +9,11 @@
         {% set latesArticle = articles.oneLates %}
         {% if latesArticle is not null %}
             <article class="main-post top-post">
-                <a href="{{ path('app_category_page', {slug: latesArticle.category.slug}) }}" class="category-link">{{ latesArticle.category.title }}</a>
-                <a href="{{ path('app_article_page', {id: latesArticle.id}) }}">
+                <a href="#" class="category-link">{{ latesArticle.categoryTitle }}</a>
+                <a href="#">
                     <img src="{{ asset(latesArticle.image) }}" alt="" class="img-fluid">
                 </a>
-                <a href="{{ path('app_article_page', {id: latesArticle.id}) }}" class="post-link">
+                <a href="#" class="post-link">
                     <h1>{{ latesArticle.title }}</h1>
                 </a>
                 <time>{{ latesArticle.publicationDate | date('d-m-Y') }}</time>
@@ -26,11 +26,11 @@
             {% for article in articles.lates(3) %}
                 <div class="col-3">
                     <article class="post top-post">
-                        <a href="{{ path('app_category_page', {slug: article.category.slug}) }}" class="category-link">{{ article.category.title }}</a>
-                        <a href="{{ path('app_article_page', {id: article.id}) }}">
+                        <a href="#" class="category-link">{{ article.categoryTitle }}</a>
+                        <a href="#">
                             <img src="{{ asset(article.image) }}" alt="" class="img-fluid">
                         </a>
-                        <a href="{{ path('app_article_page', {id: article.id}) }}" class="post-link">
+                        <a href="#" class="post-link">
                             <h3>{{ article.title }}</h3>
                         </a>
                         <time>{{ latesArticle.publicationDate | date('d-m-Y') }}</time>
@@ -44,8 +44,8 @@
     <section class="regular-posts">
         {% for article in articles %}
             <article class="post">
-                <a href="{{ path('app_category_page', {slug: article.category.slug}) }}" class="category-link">{{ article.category.title }}</a>
-                <a href="{{ path('app_article_page', {id: article.id}) }}"><h4>{{ article.title }}</h4></a>
+                <a href="#" class="category-link">{{ article.categoryTitle }}</a>
+                <a href="#"><h4>{{ article.title }}</h4></a>
                 <time>{{ article.publicationDate | date('d-m-Y') }}</time>
             </article>
         {% endfor %}
